Upgrading to React 19 to use useActionState
