+++ app/register/page.tsx
@@ -1,23 +1,13 @@
+"use client";
+
 import Link from "next/link";
 import { Form } from "app/form";
-import { redirect } from "next/navigation";
-import { createUser, getUser } from "app/db";
 import { SubmitButton } from "app/submit-button";
+import { useActionState } from "react";
+import { registerUser } from "../lib/actions";
 
 export default function Login() {
-  async function register(formData: FormData) {
-    "use server";
-    let email = formData.get("email") as string;
-    let password = formData.get("password") as string;
-    let user = await getUser(email);
-
-    if (user.length > 0) {
-      return "User already exists"; // TODO: Handle errors with useFormStatus
-    } else {
-      await createUser(email, password);
-      redirect("/login");
-    }
-  }
+  const [errorMessage, formAction] = useActionState(registerUser, undefined);
 
   return (
     <div className="flex h-screen items-center justify-center bg-gray-50 dark:bg-black px-4">
@@ -28,7 +18,7 @@ export default function Login() {
             Create an account with your email and password
           </p>
         </div>
-        <Form action={register}>
+        <Form action={formAction}>
           <SubmitButton>Sign Up</SubmitButton>
           <p className="text-center text-sm text-gray-600 dark:text-slate-400">
             {"Already have an account? "}
@@ -40,6 +30,7 @@ export default function Login() {
             </Link>
             {" instead."}
           </p>
+          <div className="text-red-700">{errorMessage}</div>
         </Form>
       </div>
     </div>
