+++ package.json
@@ -13,11 +13,11 @@
     "bcrypt-ts": "^5.0.0",
     "drizzle-orm": "^0.29.2",
     "geist": "^1.2.0",
-    "next": "^14.0.4",
+    "next": "15.1.4",
     "next-auth": "5.0.0-beta.4",
     "postgres": "^3.4.3",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "19.0.0-rc-f38c22b244-20240704",
+    "react-dom": "19.0.0-rc-f38c22b244-20240704",
     "@fortawesome/fontawesome-svg-core": "^6.7.0",
     "@fortawesome/free-brands-svg-icons": "^6.7.0",
     "@fortawesome/free-regular-svg-icons": "^6.7.0",
@@ -30,6 +30,8 @@
     "eslint-config-next": "^14.0.4",
     "postcss": "^8.4.32",
     "tailwindcss": "^3.4.0",
-    "typescript": "^5.3.3"
+    "typescript": "^5.3.3",
+    "@types/react": "18.3.3",
+    "@types/react-dom": "18.3.0"
   }
 }
